Remove StorageProvider in favor of better useStorageState
