+++ projects/bp-gallery/src/components/common/clipboard/ClipboardEditor.tsx
@@ -3,7 +3,8 @@ import { Badge, Button } from '@mui/material';
 import { difference } from 'lodash';
 import { useCallback, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { useClipboard, useClipboardEditorButtons } from '../../../hooks/context-hooks';
+import { useClipboard } from '../../../hooks/clipboard.hook';
+import { useClipboardEditorButtons } from '../../../hooks/context-hooks';
 import { FlatPicture } from '../../../types/additionalFlatTypes';
 import { AuthRole, useAuth } from '../../provider/AuthProvider';
 import { ScrollProvider } from '../../provider/ScrollProvider';
