+++ projects/bp-gallery/src/components/views/picture/sidebar/picture-info/LinkedInfoField.tsx
@@ -5,7 +5,8 @@ import { useCallback, useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useGetMultiplePictureInfoLazyQuery } from '../../../../../graphql/APIConnector';
 import { useSimplifiedQueryResponseData } from '../../../../../graphql/queryUtils';
-import { useClipboard, useSetClipboardEditorButtons } from '../../../../../hooks/context-hooks';
+import { useClipboard } from '../../../../../hooks/clipboard.hook';
+import { useSetClipboardEditorButtons } from '../../../../../hooks/context-hooks';
 import { FlatPicture } from '../../../../../types/additionalFlatTypes';
 import CheckboxButton from '../../../../common/CheckboxButton';
 import { HelpTooltip } from '../../../../common/HelpTooltip';
