+++ projects/bp-gallery/src/hooks/anonymous-id.hook.ts
@@ -1,0 +1,7 @@
+import { nanoid } from 'nanoid';
+import useStorageState from './storage-state.hook';
+
+export const useAnonymousId = () => {
+  const [anonymousId] = useStorageState<string>(() => nanoid(), 'anonymous_id', sessionStorage);
+  return anonymousId;
+};
