+++ projects/bp-gallery/src/hooks/clipboard.hook.ts
@@ -1,0 +1,9 @@
+import useStorageState from './storage-state.hook';
+
+type ClipboardData = {
+  pictureIds: string[];
+};
+
+export const useClipboard = () => {
+  return useStorageState<ClipboardData>({ pictureIds: [] }, 'clipboard', sessionStorage);
+};
